tier2_requirements.txt
opencv-python==4.12.0.88
numpy==2.0.2
torch==2.8.0
torchvision==0.19.0
realesrgan==0.3.0
gfpgan==1.3.8
scikit-image==0.25.2
Pillow==11.3.0
yt-dlp==2025.9.5
tqdm==4.67.1
scipy==1.16.2
requests==2.32.3
ffmpeg-python==0.2.0
imageio==2.37.0
imageio-ffmpeg==0.5.1
